EntityReferenceAutocomplete schema is required #225.
