+++ src/components/05_pages/NodeForm/SchemaPropType.js
@@ -1,0 +1,14 @@
+import PropTypes from 'prop-types';
+
+const lazyFunction = f => (props, propName, componentName, ...rest) =>
+  f(props, propName, componentName, ...rest);
+
+var schemaType;
+const lazySchemaType = lazyFunction(() => schemaType);
+
+export default PropTypes.shape({
+  type: PropTypes.string.isRequired,
+  title: PropTypes.string,
+  description: PropTypes.string,
+  properties: PropTypes.objectOf(lazyFunction(lazySchemaType)),
+});
